Move lightsaber proc to own file
